feat: 스탭 1,2,3,4 컴포넌트 분리
